+++ src/tyconf/__init__.py
@@ -6,9 +6,10 @@ configuration with runtime type validation and an intuitive API.
 """
 
 from .core import TyConf, PropertyDescriptor
+from . import validators
 
-__all__ = ["TyConf", "PropertyDescriptor"]
-__version__ = "1.0.2"
+__all__ = ["TyConf", "PropertyDescriptor", "validators"]
+__version__ = "1.1.0"
 __date__ = "2025-11-18"
 __author__ = "barabasz"
 __license__ = "MIT"