+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Refactored `remove()` and `__delitem__()` to use shared `_del_property()` helper method
+- Hint in get() method
 
 ## [1.0.1] - 2025-11-18